pyproject.toml
[tool.poetry]
name = "ab-trade-assignment"
version = "0.1.0"
description = ""
authors = ["Sergey Kim <<sergkim7@gmail.com>>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.96.0"
uvicorn = "^0.22.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
